refactor(navigation): split navigators out of App.tsx
Move BottomTabNavigator, ShopNavigator, CartNavigator, and DebugNavigator
into their own files under src/navigators/. App.tsx now just imports and
renders RootTabNavigator, shrinking ~100 lines of nested navigator
declarations out of the root component file.

Also drop the createNativeStackNavigator-as-createStackNavigator alias in
favor of using createNativeStackNavigator directly.
