+++ src/navigators/ShopNavigator.tsx
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+
+import HomeScreen from '../screens/HomeScreen';
+import ProductDetailScreen from '../screens/ProductDetailScreen';
+import {RootStackParamList} from '../navigation';
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+const ShopNavigator = () => {
+  return (
+    <Stack.Navigator
+      screenOptions={{
+        headerShown: false,
+      }}>
+      <Stack.Screen name="Home" component={HomeScreen} />
+      <Stack.Screen name="ProductDetail" component={ProductDetailScreen} />
+    </Stack.Navigator>
+  );
+};
+
+export default ShopNavigator;
